% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONGrazon.R
\name{cong.razon}
\alias{cong.razon}
\title{Función cong.razon}
\usage{
cong.razon(numerador, denominador, N, M, V, v = 0, alfa = 0.05)
}
\arguments{
\item{numerador:}{numerador: vector numérico de las observaciones realizadas para la variable, X.}

\item{denominador:}{vector numérico de las observaciones realizadas para la variable, Y.}

\item{V:}{vector que contiene los tamaños poblacionales de los conglomerados muestreados.}

\item{v:}{vector que contiene los tamaños de las muestras extraídas de cadaconglomerado. (v = 0, representará un muestreo por conglomerados
unietápico).}

\item{N:}{tamaño de la población muestreada.}

\item{M:}{número de conglomerados en los que se divide la población.}

\item{alfa:}{nivel de error (1-alfa: nivel de confianza).}
}
\value{
Estudio Inferencial para el parámetro razón poblacional a partir
de un muestreo por conglomerados unietápico (v = 0; CON1) ó bietápico (CON2):

-Estimación: estimación del parámetro razón (estimador: razón
                                              ponderada muestral).

 -Varianza: estimación de la varianza de la razón ponderada muestral.

 -Error de muestreo: estimación de la desviación típica de la razón ponderada muestral.

 -Intervalos de confianza: intervalos de confianza al nivel 1-alfa
 para el parámetro razón poblacional, bajo los supuestos
 de normalidad y no normalidad (s.n.-sin normalidad) en la población.
}
\description{
La función cong.razon es utilizada para estudiar el parámetro razón
de los totales poblacionales de X respecto a Y, cuando se realiza un
muestreo por conglomerados (unietápico o bietápico).
}
\examples{
cong.razon(x<-c(rnorm(1000,50,3),rnorm(2000,50,3),rnorm(500,50,3)),
y<-c(rnorm(1000,75,4),rnorm(2000,75,4),rnorm(500,75,4)),
V=c(1000,2000,500), N=35000, M=10)
}
