% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONGtotal.R
\name{cong.total}
\alias{cong.total}
\title{Función cong.total}
\usage{
cong.total(y, N, M, V, v = 0, d = 5, alfa = 0.05)
}
\arguments{
\item{y:}{vector numérico de las observaciones realizadas.}

\item{V:}{vector que contiene los tamaños poblacionales de los conglomerados muestreados.}

\item{v:}{vector que contiene los tamaños de las muestras extraídas de cada conglomerado. (v = 0, representará un muestreo por conglomerados unietápico).}

\item{N:}{tamaño de la población muestreada.}

\item{M:}{número de conglomerados en los que se divide la población.}

\item{d:}{número de intervalos o clases utilizados para  construir el histograma de los datos observados y la tabla de frecuencias, cuando corresponda.}

\item{alfa:}{nivel de error (1-alfa: nivel de confianza).}
}
\value{
Estudio Inferencial para el parámetro total poblacional a partir de un
muestreo por conglomerados unietápico (v = 0; CON1) ó bietápico (CON2):

 -Estimación: estimación del parámetro total (estimador: suma
                                              ponderada muestral).

-Varianza: estimación de la varianza de la suma ponderada muestral.

-Error de muestreo: estimación de la desviación típica de la suma ponderada muestral.

-Intervalos de confianza: intervalos de confianza al nivel 1-alfa
para el parámetro total poblacional, bajo los supuestos de normalidad y
no normalidad (s.n.-sin normalidad) en la población.

Estudio Descriptivo:

-Gráfico: Histograma de frecuencias absolutas y estimación de la función de densidad.

-Estadísticos muestrales: Tamaño, media, mediana, varianza, desviación típica, coeficiente de variación, rango
intercuartílico, coeficiente de asimetría y coeficiente de curtosis.

-Tabulación: Tabla de frecuencias (muestras mayores de 50) ó diagrama tallo-hoja (en otro caso).
}
\description{
La función cong.total es utilizada para estudiar el parámetro total
de una variable cuando se realiza un muestreo por conglomerados (unietápico o bietápico).
}
\examples{
cong.total(x<-c(rnorm(1000,50,3),rnorm(2000,50,3),rnorm(500,50,3)),
V=c(1000,2000,500), N=35000, M=10, d=20)
}
