% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MASmedia.R
\name{mas.media}
\alias{mas.media}
\title{Función mas.media}
\usage{
mas.media(y, N, d = 5, alfa = 0.05, delta = 0.5)
}
\arguments{
\item{y:}{vector numérico de las observaciones realizadas.}

\item{N:}{tamaño de la población muestreada.}

\item{d:}{número de intervalos o clases utilizados para construir el histograma de los datos observados y la tabla de frecuencias, cuando corresponda.}

\item{alfa:}{nivel de error (1-alfa: nivel de confianza).}

\item{delta:}{error relativo.}
}
\value{
Estudio Inferencial para el parámetro media poblacional a partir de un muestreo aleatorio simple (M.A.S.):

-Estimación: estimación del parámetro media (estimador: media muestral).

-Varianza: estimación de la varianza de la media muestral.

-Error de muestreo: estimación de la desviación típica de la media muestral.

          -Intervalos de confianza: intervalos de confianza al nivel 1-alfa para el parámetro media poblacional,
          bajo los supuestos de normalidad y no normalidad (s.n.-sin normalidad) en la población.

-Tamaño muestral: tamaño muestral necesario para cometer un error relativo delta para un nivel de confianza 1-alfa,
 bajo los supuestos de normalidad y no normalidad (s.n.-sin normalidad) en la población.



Estudio Descriptivo:

-Gráfico: Histograma de frecuencias absolutas y estimación de la
función de densidad.

-Estadísticos muestrales: Tamaño, media, mediana, varianza,
desviación típica, coeficiente de variación, rango
intercuartílico, coeficiente de asimetría y coeficiente
de curtosis.

 -Tabulación: Tabla de frecuencias (muestras mayores de 50) ó diagrama tallo-hoja (en otro caso).
}
\description{
La función mas.media es utilizada para estudiar el parámetro media de una variable cuando se realiza un muestreo aleatorio simple.
También es válida para proporciones utilizando variables dicotómicas.
}
\examples{
mas.media(rnorm(1000,75,3), N=10000, d=20, delta=0.001)

}
