% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MASrazon.R
\name{mas.razon}
\alias{mas.razon}
\title{Función mas.razon}
\usage{
mas.razon(numerador, denominador, N, alfa = 0.05)
}
\arguments{
\item{numerador:}{vector numérico de las observaciones realizadas para la variable X.}

\item{denominador:}{vector numérico de las observaciones realizadas para la variable Y.}

\item{N:}{tamaño de la población muestreada.}

\item{alfa:}{nivel de error (1-alfa: nivel de confianza).}
}
\value{
Estudio Inferencial para el parámetro razón poblacional a partir de un muestreo aleatorio simple (M.A.S.):

-Estimación: estimación del parámetro razón (estimador: razón muestral).

-Varianza: estimación de la varianza de la razón muestral.

-Error de muestreo: estimación de la desviación típica de la razón muestral.

          -Intervalos de confianza: intervalos de confianza al nivel 1-alfa para el parámetro razón poblacional,
          bajo los supuestos de normalidad y no normalidad (s.n.-sin normalidad) en la población.
}
\description{
La función mas.razon es utilizada para estudiar el parámetro razón de los totales poblacionales de X respecto a Y, cuando se realiza un muestreo aleatorio simple.
}
\examples{
mas.razon(rnorm(1000,75,3), rnorm(1000,54,1), N=10000, alfa=0.01)
}
