% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSmedia.R
\name{ms.media}
\alias{ms.media}
\title{Función ms.media}
\usage{
ms.media(y, N, k, d = 5, alfa = 0.05)
}
\arguments{
\item{y:}{vector numérico de las observaciones realizadas.}

\item{N:}{tamaño de la población muestreada.}

\item{k:}{paso utilizado para extraer la muestra (N=k*n, siendo n el tamaño muestral).}

\item{d:}{número de intervalos o clases utilizados para construir el histograma de los datos observados y la tabla de frecuencias, cuando corresponda.}

\item{alfa:}{nivel de error (1-alfa: nivel de confianza).}
}
\value{
Estudio Inferencial para el parámetro media poblacional a partir de un muestreo sistemático:

-Estimación: estimación del parámetro media (estimador: media muestral).

-Varianza: estimación de la varianza de la media muestral. Método Jackknife.

-Error de muestreo: estimación de la desviación típica de la media muestral.

-Intervalos de confianza: intervalos de confianza al nivel 1-alfapara el parámetro media poblacional, bajo los supuestos
de normalidad y no normalidad (s.n.-sin normalidad) en la población.

Estudio Descriptivo:

-Gráfico: Histograma de frecuencias absolutas y estimación de la función de densidad.

-Estadísticos muestrales: Tamaño, media, mediana, varianza, desviación típica, coeficiente de variación, rango
intercuartílico, coeficiente de asimetría y coeficiente de curtosis.

-Tabulación: Tabla de frecuencias (muestras mayores de 50) ó diagrama tallo-hoja (en otro caso).
}
\description{
La función ms.media es utilizada para estudiar el parámetro media de una variable cuando se realiza un muestreo sistemático. También
es válida para proporciones utilizando variables dicotómicas.
}
\examples{
ms.media(rnorm(1000,75,3), N=10000, k=10, d=20)

}
