% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MStotal.R
\name{ms.total}
\alias{ms.total}
\title{Función ms.total}
\usage{
ms.total(y, N, k, d = 5, alfa = 0.05)
}
\arguments{
\item{y:}{vector numérico de las observaciones realizadas.}

\item{N:}{tamaño de la población muestreada.}

\item{k:}{paso utilizado para extraer la muestra (N=k*n, siendo n el tamaño muestral).}

\item{d:}{número de intervalos o clases utilizados para construir el histograma de los datos observados y la tabla de frecuencias, cuando corresponda.}

\item{alfa:}{nivel de error (1-alfa: nivel de confianza).}
}
\value{
Estudio Inferencial para el parámetro total poblacional a partir de un muestreo sistemático:

    -Estimación: estimación del parámetro total (estimador: suma muestral).

    -Varianza: estimación de la varianza de la suma muestral. Método Jackknife.

    -Error de muestreo: estimación de la desviación típica de la suma muestral.

    -Intervalos de confianza: intervalos de confianza al nivel 1-alfa para el parámetro total poblacional, bajo los supuestos
    de normalidad y no normalidad (s.n.-sin normalidad) en la población.


    Estudio Descriptivo:

    -Gráfico: Histograma de frecuencias absolutas y estimación de la función de densidad.

    -Estadísticos muestrales: Tamaño, media, mediana, varianza, desviación típica, coeficiente de variación, rango
    intercuartílico, coeficiente de asimetría y coeficiente de curtosis.

    -Tabulación: Tabla de frecuencias (muestras mayores de 50) ó diagrama tallo-hoja (en otro caso).
}
\description{
La función ms.total es utilizada para estudiar el parámetro total
una variable cuando se realiza un muestreo aleatorio simple.
}
\examples{
ms.total(rnorm(1000,75,3), N=10000, k=10, d=20)
}
